Abstract commands signature rendering component
