+++ apps/website/src/components/CopyToClipboardButton.tsx
@@ -26,12 +26,12 @@ const CopyToClipboardButton = ({
   return (
     <div className="group relative inline-block">
       <button onClick={onClick} title="Copy to clipboard">
-        <IconClipboard className="ml-2 inline size-4 cursor-pointer text-alveus-green-400 duration-300 hover:text-black" />
+        <IconClipboard className="m-1 inline size-4 cursor-pointer text-alveus-green-400 duration-300 hover:text-black" />
       </button>
 
       <span
         className={classes(
-          "pointer-events-none absolute top-0 left-full ml-2 rounded-md bg-alveus-green-900 px-1 text-white transition-opacity duration-300",
+          "pointer-events-none absolute top-1/2 left-full -translate-y-1/2 rounded-md bg-alveus-green-900 px-1 text-white transition-opacity duration-300",
           status === undefined ? "opacity-0" : "opacity-100",
           "group-hover:opacity-100",
         )}
